Include SequenceReaderExtensions2 in DynamicCodeDumper.csproj
Code cleanup

* Avoid using Utilities extension in BufferWriter
* Move NilFormatter to MessagePack.Formatters
